+++ tests/testthat/test-utils.R
@@ -54,3 +54,121 @@ test_that("getSelector", {
   expect_equal(getSelector("mykey", "%", exact = TRUE), " AND mykey LIKE '%'")
   expect_equal(getSelector("mykey", "%", exact = FALSE), " AND mykey LIKE '%'")
 })
+
+test_that("configure_spark() returns early on windows", {
+  local_mocked_bindings(is_macos = function() {FALSE})
+
+  expect_equal(configure_spark(), NULL)
+})
+
+test_that("configure_spark() errors informatively on failure to install unixODBC", {
+  local_mocked_bindings(
+    is_macos = function() {TRUE},
+    locate_install_unixodbc = function() {character(0)}
+  )
+
+  expect_snapshot(databricks(), error = TRUE)
+})
+
+test_that("locate_install_unixodbc() returns reasonable values", {
+  skip_if(!is_macos())
+  skip_if(!has_unixodbc(), "odbcinst not available.")
+
+  res <- locate_install_unixodbc()
+
+  expect_true(file.exists(res[1]))
+  expect_true(grepl("\\.dylib", res[1]))
+})
+
+test_that("databricks() errors informatively when spark ini isn't writeable", {
+  local_mocked_bindings(is_writeable = function(path) {FALSE})
+  expect_snapshot(
+    write_spark_lines("", ".", ".", call2("databricks")),
+    error = TRUE
+  )
+})
+
+test_that("locate_config_spark() returns reasonable values", {
+  simba_spark_ini <- "some/folder/simba.sparkodbc.ini"
+  withr::local_envvar(SIMBASPARKINI = simba_spark_ini)
+  expect_equal(locate_config_spark(), simba_spark_ini)
+})
+
+test_that("configure_unixodbc_spark() writes reasonable entries", {
+  unixodbc_install_path <- "libodbcinst.dylib"
+  spark_config_path <- "simba.sparkodbc.ini"
+
+  withr::local_file(spark_config_path)
+
+  # neither of the relevant fields already there:
+  writeLines(
+    c("some=entries", "not=relevant"),
+    con = spark_config_path
+  )
+
+  configure_unixodbc_spark(
+    unixodbc_install = unixodbc_install_path,
+    spark_config = spark_config_path
+  )
+
+  expect_equal(
+    readLines(spark_config_path),
+    c(
+      "some=entries",
+      "not=relevant",
+      "ODBCInstLib=libodbcinst.dylib",
+      "DriverManagerEncoding=UTF-16"
+    )
+  )
+
+  # both of the relevant fields are already there:
+  writeLines(
+    c("some=entries",
+      "not=relevant",
+      "ODBCInstLib=somewhere.dylib",
+      "DriverManagerEncoding=UTF-8"),
+    con = spark_config_path
+  )
+
+  res <- configure_unixodbc_spark(
+    unixodbc_install = unixodbc_install_path,
+    spark_config = spark_config_path
+  )
+
+  expect_equal(res, NULL)
+  expect_equal(
+    readLines(spark_config_path),
+    c(
+      "some=entries",
+      "not=relevant",
+      "ODBCInstLib=libodbcinst.dylib",
+      "DriverManagerEncoding=UTF-16"
+    )
+  )
+
+  # an entry is there but commented out
+  writeLines(
+    c("some=entries",
+      "not=relevant",
+      ";ODBCInstLib=somewhere.dylib",
+      ";DriverManagerEncoding=UTF-8"),
+    con = spark_config_path
+  )
+
+  configure_unixodbc_spark(
+    unixodbc_install = unixodbc_install_path,
+    spark_config = spark_config_path
+  )
+
+  expect_equal(
+    readLines(spark_config_path),
+    c(
+      "some=entries",
+      "not=relevant",
+      ";ODBCInstLib=somewhere.dylib",
+      ";DriverManagerEncoding=UTF-8",
+      "ODBCInstLib=libodbcinst.dylib",
+      "DriverManagerEncoding=UTF-16"
+    )
+  )
+})