+++ NEWS.md
@@ -1,5 +1,8 @@
 # odbc (development version)
 
+* `databricks()` will now automatically configure the needed driver and driver
+  manager on macOS (#651).
+
 * Transitioned `odbcDataType()` to use S4 for consistency. S3 methods defined
   locally will need to be rewritten (#701).